Refactor: Moves LogLabels and LogLabel to grafana/ui
